Improved account set handling
We iterated over the length of the account set but account was picked randomly. Worst case scenario would be the same account every iteration.
To fix this random shuffle of all items in the list was applied. This way we can be certain to randomly iterate all accounts and yet only test each account just once.